% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarity.R
\name{automatic.correspondences}
\alias{automatic.correspondences}
\title{Automatic correspondences}
\usage{
automatic.correspondences(contour, R, plot = F, cex = 2)
}
\arguments{
\item{contour}{data frame, a polygon data frame with list items x, y, and id.}

\item{R}{integer, parameter setting coarseness of generated homology points.}

\item{plot}{boolean, if the result should be displayed.}

\item{cex}{numeric, size of plotted points.}
}
\description{
Generate homology points from a contour.
}
\examples{
nuclei.folder<-system.file('roi/RoiSet_nuclei', package='leeplyr')
contour<-get.rois(nuclei.folder)
contour<-contour[which(contour$id==3),1:2]
corr.points<-automatic.correspondences(contour, R = 4)
}
\keyword{homology}
\keyword{points}
