% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roianalysis.R
\name{get.rois}
\alias{get.rois}
\title{Get ROIs}
\usage{
get.rois(folder)
}
\arguments{
\item{folder}{character describing the system file path for the folder containing ROIs (can't be a zip-file) so unzip before.}
}
\description{
This function reads in a set of ImageJ ROIs and returns a polygon data frame.
}
\examples{
cytosol.folder<-system.file('data/roi/RoiSet_cytosol', package='leeplyr')
nuclei.folder<-system.file('data/roi/RoiSet_cytosol', package='leeplyr')
cytosol<-get.rois(cytosol.folder)
nuclei<-get.rois(nuclei.folder)
}
\keyword{ROI}
