% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarity.R
\name{map.to.polarity}
\alias{map.to.polarity}
\title{Map polarity}
\usage{
map.to.polarity(transcripts, contour, plot = TRUE)
}
\arguments{
\item{transcripts}{FISSEQ readr output tsv file path or a tibble object or data frame.}

\item{contour}{data frame, a polygon data frame with list items x, y, and id.}

\item{plot}{boolean, if the result should be displayed.}
}
\description{
Uses PCA to map polarity of transcripts given an ROI.
}
\examples{
nuclei.folder<-system.file('roi/RoiSet_nuclei', package='leeplyr')
contour<-get.rois(nuclei.folder)
transcripts<-system.file('fisseq/res_001_001FISSEQ.out', package='leeplyr')
polarity<-map.to.polarity(transcripts, contour, plot = TRUE)
}
\keyword{polarity}
