% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roianalysis.R
\name{map.to.roi}
\alias{map.to.roi}
\title{Map transcripts to ROI}
\usage{
map.to.roi(transcripts, roi.folder = c(cytosol.folder, nuclei.folder),
  roi.labels = c("cytoplasm", "nucleus"), OR.cutoff = 5,
  p.value.cutoff = 20, plot = TRUE, roi.col = c("#C8EAEB", "#FCDAD1"),
  transcript.col = "black", cex = 0.2)
}
\arguments{
\item{transcripts}{FISSEQ readr output tsv file path.}

\item{roi.folder}{character vector, each element is the location of folder containing ImageJ roi files.}

\item{roi.labels}{character vector, .}

\item{OR.cutoff}{numeric, absolute OR value on log2 coordinates to use as cut-off. Default is 5.}

\item{p.value.cutoff}{numeric, p-value on negative log10 coordinates to use as cut-off. Default is 20.}

\item{plot}{boolean, plot output graph. Default is TRUE.}

\item{roi.col}{character vector, Color of regions of interest. Default is c('#C8EAEB', '#FCDAD1').}

\item{transcript.col}{character, color or transcripts in plot. Default is 'black'.}

\item{cex}{numeric, Size of transcripts on plot showing alll transcripts. Default is 0.2.}
}
\description{
Maps each individual transcripts to one or more region of interests (ROIs).
}
\examples{
cytosol.folder<-system.file('roi/RoiSet_cytosol', package='leeplyr')
nuclei.folder<-system.file('roi/RoiSet_nuclei', package='leeplyr')
transcripts<-system.file('fisseq/res_001_001FISSEQ.out', package='leeplyr')
map.to.roi(transcripts, roi.folder = c(cytosol.folder, nuclei.folder), roi.labels =  c('cytoplasm', 'nucleus'), OR.cutoff = 5, p.value.cutoff = 20)
}
