% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roianalysis.R
\name{plot.polygon}
\alias{plot.polygon}
\title{Plot polygon}
\usage{
plot.polygon(dataframe, col)
}
\arguments{
\item{dataframe}{data frame, a polygon data frame with list items x, y, and id.}

\item{col}{character, fill color for polygon. character}
}
\description{
Draws polygon for a polygon data frame.
}
\examples{
cytosol<-get.rois(cytosol.folder)
plot(cytosol[,1:2], col = 0)
plot.polygon(cytosol, col='pink')
}
\keyword{polygon}
