% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/administration.R
\name{update.leeplyr}
\alias{update.leeplyr}
\title{Update leeplyr package}
\usage{
\method{update}{leeplyr}()
}
\description{
This function updates the leeplyr packae.
}
\examples{
update.leeplyr
}
\keyword{update}
